\name{angle}
\alias{angle}
\title{angle}
\description{A function for calculating bond angle and 1-3 distance for any pdb file.}
\usage{angle(id,pdb)}
\details{
An internal function, to do calculations of virtual angles and virtual 1-3 distances for an uploaded protein structure data. 
}
\arguments{
\item{id}{a character string to describe the protein name, sometimes of length 4 and refer the protein ID as shown in PDB }
\item{pdb}{a .pdb file downloaed from PDB or generated by users}
}
\value{

\item{Angle}{ a numeric matrix including the columns for the computed virtual angles, res1-res3 distance, and the columns for the corresponding residue name}
}
\author{Yuanyuan Huang, Stephen Bonett, and Zhijun Wu}
\examples{
id<-"1ABA"
pdb<-read.pdb(id)
angle(id,pdb)
}
\keyword{internal function virtual bond angle}
