\name{bond}
\alias{bond}
\title{bond}
\description{A function for calculating virtual bond length between two neighbouring residues of a specific pdb file}
\usage{bond(id,pdb)}
\details{
An internal function, to do calculations of virtual bond length for an uploaded protein structure data. 
}
\arguments{
\item{id}{a character string to describe the protein name, sometimes of length 4 and refer the protein ID as shown in PDB }
\item{pdb}{a .pdb file downloaed from PDB or generated by users}
}
\value{

\item{Bond}{ a numeric matrix including the columns for the computed virtual bond length and the columns for the corresponding residue name}
}

\examples{
id<-"1ABA"
pdb<-read.pdb(id)
bond(id,pdb)
}
\author{Yuanyuan Huang, Stephen Bonett, and Zhijun Wu}
\keyword{internal function virtual bond length}