\name{callBond}
\alias{callBond}
\title{
callBond
}
\description{
Opens the dialogue for bond angle queries. This can display various graphics related to 1-2 bond distances.
}
\details{
An internal function, usually called by pressing the "Bond Distance - Two Residues" button. Query the database by selecting a series of two residues. Choosing "Any" for one or more of the residues will sum over all possible values of those residues. Choosing "Any" for both residues will produce the full histogram. Finally, bin sizes can be changes to make the histograms readable, and there is an option to display either raw frequencies or density on the y-axis.
}
\author{Yuanyuan Huang, Stephen Bonett, and Zhijun Wu}
\keyword{internal}
