\name{callEval}
\alias{callEval}
\title{
callEval
}
\description{
Opens the dialogue for evaluating a user supplied protein. Can evalate statistical potentials and produce ramachandran plots.
}
\details{
An internal function that is usually called by pressing the "Structural Analysis" button. Enter a convient name for your protein for reference and either enter the path to your .pdb file or use the browse button to browse for it. Once protein is selected and names, either distance or angle potentials can be calculated. Ramachandran plots can also be produced. 
}
\author{Yuanyuan Huang, Stephen Bonett, and Zhijun Wu}
\keyword{internal}
