\name{data ata}
\alias{data.ata}
\title{
data.ata}
\description{
A function to combind the virtual bond angle, and virtual torsion angle files of one protein into a data set so that get an alpha-tau-beta matrix.
}
\usage{data.ata(id,pdb)}
\details{
An internal function, to do calculations of virtual angles and virtual torsional angles for a protein structure data. 
}
\arguments{
\item{id}{a character string to describe the protein name, sometimes of length 4 and refer the protein ID as shown in PDB }
\item{pdb}{a .pdb file downloaed from PDB or generated by users}
}
\value{\item{ata}{ a numeric matrix including the columns for the computed 1st virtual angles alpha, virtual torsional angles tau, 2nd virtual bond angles beta, and the columns for the corresponding residue name}}
\examples{
id<-"1ABA"
pdb<-read.pdb(id)
data.ata(id,pdb)
}
\author{Yuanyuan Huang, Stephen Bonett, and Zhijun Wu}
\keyword{internal funtion combine alpha-tau-beta}
