\name{press-package}
\alias{press-package}
\docType{package}
\title{
Protein REsidue-Level Structural Statistics
}
\description{
Tools for graphically displaying distributional information about residue-level structural properties. Also includes a method to model and analyze a user provided protein structure in terms of its structural properties.
}
\details{
\tabular{ll}{
Package: \tab press\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2011-4-26\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
The GUI environment is accessed through the function call press(). This function will open a graphical menu gives access the various functional modules in the software. Most of these modules accept residue sequence queries and display statistical distributions graphically. The "structural analysis" module is a tool for analyzing user provided structures. A tutorial is available in the help menu.
}
\author{
Yuanyuan Huang, Stephen Bonett, and Zhijun Wu

Maintainer: Yuanyuan Huang <sunnyuan.h@gmail.com>

}

\keyword{ package }

