\name{press}
\alias{press}
\title{P.R.E.S.S. Main Menu
}

\description{
Opens the main menu for the P.R.E.S.S. GUI environment. From this menu, all of the functionality of P.R.E.S.S. can be accessed.
}

\usage{press()}

\details{
Each button opens a dialogue with different functionality. Query the database by selection either "Bond Distance - Two Residues", "Bond Angle - Residues", "Torsion,ATA - Four Residues", or "TAT - Five Residues". Choose "Structure Analysis" to evaluate a novel protien. Choose "Info" for software information. 
}
\author{Yuanyuan Huang, Stephen Bonett, and Zhijun Wu}