\name{pressbond}
\alias{pressbond}
\title{pressbond}
\description{A function for the estimation of potentials of an uploaded pdb file.}
\usage{pressbond(id,pdb)}
\details{
An internal function, to do calculations of virtual bond length and display the dynamics of bonds potentials for an uploaded protein structure data. 
}
\arguments{
\item{id}{a character string to describe the protein name, sometimes of length 4 and refer the protein ID as shown in PDB }
\item{pdb}{a .pdb file downloaed from PDB or generated by users}
}
\author{Yuanyuan Huang, Stephen Bonett, and Zhijun Wu}
\keyword{internal function bonds potentials}
