\name{presscont1}
\alias{presscont1}
\title{presscont1}
\description{A function for the ramachandran-like plot of alpha-tau for an uploaded pdb file.}
\usage{presscont1(id,pdb)}
\details{
An internal function, to do calculations of virtual bond angles, virtual torsional angles and display the alpha-tau correlations with ramachandran-like plot for an uploaded protein structure data. 
}
\arguments{
\item{id}{a character string to describe the protein name, sometimes of length 4 and refer the protein ID as shown in PDB }
\item{pdb}{a .pdb file downloaed from PDB or generated by users}
}
\author{Yuanyuan Huang, Stephen Bonett, and Zhijun Wu}
\keyword{internal function alpha-tau plot}