\name{presscont2}
\alias{presscont2}
\title{presscont2}
\description{A function for the ramachandran-like plot of beta-tau for an uploaded pdb file.}
\usage{presscont2(id,pdb)}
\details{
An internal function, to do calculations of virtual bond angles, virtual torsional angles and display the beta-tau correlations with ramachandran-like plot for an uploaded protein structure data. 
}
\arguments{
\item{id}{a character string to describe the protein name, sometimes of length 4 and refer the protein ID as shown in PDB }
\item{pdb}{a .pdb file downloaed from PDB or generated by users}
}
\author{Yuanyuan Huang, Stephen Bonett, and Zhijun Wu}
\keyword{internal function beta-tau plot}
    
