\name{torsion}
\alias{torsion}
\title{torsion}
\description{A function for computing all virtual torsional angles for any pdb file.}
\usage{torsion(id,pdb)}
\details{
An internal function, to do calculations of virtual torsional angles and virtual 1-4 distances for an uploaded protein structure data. 
}
\arguments{
\item{id}{a character string to describe the protein name, sometimes of length 4 and refer the protein ID as shown in PDB }
\item{pdb}{a .pdb file downloaed from PDB or generated by users}
}
\value{

\item{Torsion}{ a numeric matrix including the columns for the computed virtual torsional angles, res1-res4 distance, and the columns for the corresponding residue name}
}
\examples{
id<-"1ABA"
pdb<-read.pdb(id)
torsion(id,pdb)
}
\author{Yuanyuan Huang, Stephen Bonett, and Zhijun Wu}
\keyword{internal function virtual torsional angle}
    